pyproject.toml
[tool.poetry]
name = "powermeter"
version = "0.1.0"
description = ""
authors = ["Tomás Rosales <trosales@mercadolibre.com>"]

[tool.poetry.dependencies]
python = "^3.8.0"
django = "==4.0.8"
django-extensions = "3.2.3"
django-environ = "==0.9.0"
django-model-utils = "==4.3.1"
djangorestframework = "==3.14.0"
django-cors-headers = "==3.13.0"
werkzeug = {version = "==2.2.2", extras = ["watchdog"]}
whitenoise = "==6.3.0"
argon2-cffi = "==21.3.0"
psycopg2-binary = "^2.9.6"
ipython = "8.12.2"
graphene-django = "^3.1.5"
Faker = "^19.9.0"
django-graphql-jwt = "^0.4.0"
django-filter = "^23.3"

[tool.poetry.dev-dependencies]
ipdb = "==0.13.11"
flake8 = "*"
flake8-isort = "==6.0.0"
black = "==22.12.0"
pylint-django = "==2.5.3"
django-debug-toolbar = "==3.8.1"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
faker = "^19.1.0"
